+++ GameApplication/GameApplication/Views/Home/Index.cshtml
@@ -2,6 +2,12 @@
     ViewData["Title"] = "Home Page";
 }
 
+<script>
+    $(document).ready(function () {
+        startChat();
+    });
+</script>
+
 <!-- TODO move to a separate partial -->
 <!-- Chatroom snippet (signalR sockets) -->
 <h2>SignalR sockets by miczi</h2>
