+++ GameApplication/GameApplication/Views/Shared/_Layout.cshtml
@@ -11,6 +11,17 @@
         <link rel="stylesheet" href="https://fonts.googleapis.com/icon?family=Material+Icons">
         <link rel="stylesheet" href="https://code.getmdl.io/1.3.0/material.cyan-light_blue.min.css">
     </environment>
+    <!-- Development specific scripts -->
+    <environment include="Development">
+        <script src="~/lib/jquery/dist/jquery.js"></script>
+        <script src="~/lib/bootstrap/dist/js/bootstrap.js"></script>
+
+        <!-- Regular SignalR -->
+        <script src="~/lib/signalr-client/dist/browser/signalr-client-1.0.0-alpha2-final.js"></script>
+        <script src="~/js/hubs/chat.js" asp-append-version="true"></script>
+        <script src="~/js/hubs/lobby.js" asp-append-version="true"></script>
+    </environment>
+
 </head>
 <body>
     <div class="demo-layout mdl-layout mdl-js-layout mdl-layout--fixed-drawer mdl-layout--fixed-header">
@@ -87,15 +98,7 @@
         <script src="https://code.getmdl.io/1.3.0/material.min.js"></script>
     </environment>
 
-     <!-- Development specific scripts -->
-    <environment include="Development">
-        <script src="~/lib/jquery/dist/jquery.js"></script>
-        <script src="~/lib/bootstrap/dist/js/bootstrap.js"></script>
 
-        <!-- Regular SignalR -->
-        <script src="~/lib/signalr-client/dist/browser/signalr-client-1.0.0-alpha2-final.js"></script>
-        <script src="~/js/site.js" asp-append-version="true"></script>
-    </environment>
 
     <!-- Place for minified scripts, production specific -->
     <environment exclude="Development">
